+++ tests/TestCase/Command/Install/CreatePluginsLinksCommandTest.php
@@ -32,14 +32,6 @@ class CreatePluginsLinksCommandTest extends TestCase
         $this->loadPlugins(['TestPlugin']);
 
         @array_map(IS_WIN ? 'rmdir_recursive' : 'unlink', [WWW_ROOT . 'me_tools', WWW_ROOT . 'test_plugin']);
-        $this->exec('me_tools.create_plugins_links');
-        $this->assertExitWithSuccess();
-        $this->assertOutputEmpty();
-        $this->assertErrorEmpty();
-        $this->assertFileExists(WWW_ROOT . 'me_tools');
-        $this->assertFileExists(WWW_ROOT . 'test_plugin');
-
-        @array_map(IS_WIN ? 'rmdir_recursive' : 'unlink', [WWW_ROOT . 'me_tools', WWW_ROOT . 'test_plugin']);
         $this->exec('me_tools.create_plugins_links -v');
         $this->assertExitWithSuccess();
         $this->assertOutputContains('Skipping plugin `Assets`. It does not have webroot folder');
@@ -50,13 +42,6 @@ class CreatePluginsLinksCommandTest extends TestCase
         $this->assertErrorEmpty();
         $this->assertFileExists(WWW_ROOT . 'me_tools');
         $this->assertFileExists(WWW_ROOT . 'test_plugin');
-
-        //Already exist
-        $this->exec('me_tools.create_plugins_links -v');
-        $this->assertExitWithSuccess();
-        $this->assertOutputContains('File or directory `' . rtr(WWW_ROOT . 'me_tools') . '` already exists');
-        $this->assertOutputContains('File or directory `' . rtr(WWW_ROOT . 'test_plugin') . '` already exists');
-        $this->assertErrorEmpty();
 
         @array_map(IS_WIN ? 'rmdir_recursive' : 'unlink', [WWW_ROOT . 'me_tools', WWW_ROOT . 'test_plugin']);
         $this->removePlugins(['TestPlugin']);